ETL_coderequirements.txt
affine==2.4.0
arviz==0.22.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==6.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
cons==0.4.7
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
etuples==0.3.10
executing==2.2.0
fastprogress==1.0.3
filelock==3.18.0
fiona==1.10.1
fonttools==4.58.2
geopandas==1.1.0
gevent==25.5.1
git-filter-repo==2.47.0
greenlet==3.2.3
h5netcdf==1.6.4
h5py==3.14.0
haversine==2.9.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lightgbm==4.6.0
logical-unification==0.4.6
lxml==6.0.0
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
miniKanren==1.0.5
multipledispatch==1.0.0
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.22
Pygments==2.19.2
pymc==5.25.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pytensor==2.31.7
python-dateutil==2.9.0.post0
python-deriv-api==0.1.6
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
rasterio==1.4.3
rasterstats==0.20.0
reactivex==4.0.4
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
shapely==2.1.1
simplejson==3.20.2
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
Theano-PyMC==1.1.2
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.3
xarray==2025.7.1
xarray-einstats==0.9.1
zope.event==5.0
zope.interface==7.2
